requirements.txt
affine==2.4.0+computecanada
aiohappyeyeballs==2.4.4+computecanada
aiohttp==3.10.9+computecanada
aiosignal==1.3.1+computecanada
attrs==24.2.0+computecanada
certifi==2024.12.14+computecanada
charset_normalizer==3.4.0+computecanada
click==8.1.7+computecanada
click-plugins==1.1.1+computecanada
cligj==0.7.2+computecanada
docker-pycreds==0.4.0+computecanada
einops==0.8.0+computecanada
filelock==3.16.1+computecanada
frozenlist==1.5.0+computecanada
fsspec==2024.10.0+computecanada
geopandas==1.0.1+computecanada
gitdb==4.0.11+computecanada
GitPython==3.1.43+computecanada
huggingface-hub==0.27.0
idna==3.10+computecanada
imageio==2.36.1+computecanada
jinja2==3.1.4+computecanada
joblib==1.4.2+computecanada
jsonschema==4.23.0+computecanada
jsonschema_specifications==2024.10.1+computecanada
laspy==2.5.4
laszip==0.2.3
lightning==2.4.0+computecanada
lightning_utilities==0.11.8+computecanada
mamba-ssm==2.2.2
MarkupSafe==2.1.5+computecanada
msgpack==1.1.0+computecanada
multidict==6.1.0+computecanada
networkx==3.4.2+computecanada
ninja==1.11.1+computecanada
numpy==1.26.4+computecanada
Pillow_SIMD==9.5.0.post2+computecanada
pointnext==0.0.5
propcache==0.2.1+computecanada
protobuf==5.29.1+computecanada
pyogrio==0.10.0+computecanada
pyproj==3.6.1+computecanada
pytorch_lightning==2.4.0+computecanada
PyYAML==6.0.2+computecanada
rasterio==1.3.10+computecanada
ray==2.37.0+computecanada
referencing==0.35.1+computecanada
regex==2024.9.11+computecanada
requests==2.32.3+computecanada
rpds_py==0.20.0+computecanada
safetensors==0.4.5+computecanada
scikit_learn==1.5.2+computecanada
seaborn==0.13.2+computecanada
sentry_sdk==2.19.0+computecanada
setproctitle==1.3.4+computecanada
shapely==2.0.4+computecanada
smmap==5.0.1+computecanada
snuggs==1.4.7+computecanada
sympy==1.13.1+computecanada
tensorboardX==2.6.2.2+computecanada
threadpoolctl==3.5.0+computecanada
tokenizers==0.21.0+computecanada
torch==2.4.1+computecanada
torch_scatter==2.1.2+torch24.computecanada
torchaudio==2.4.1+computecanada
torchdata==0.8.0+computecanada
torcheval==0.0.7+computecanada
torchmetrics==1.5.2+computecanada
torchsummary==1.5.1+computecanada
torchtext==0.18.0+computecanada
torchvision==0.19.1+computecanada
tqdm==4.67.1+computecanada
transformers==4.47.1
triton==3.1.0+computecanada
urllib3==2.2.3+computecanada
wandb==0.18.0+computecanada
yarl==1.18.0+computecanada
